recon: stateDB recovery after a peer crash
Once the pvtData of old blocks are committed to pvtStore,
we need to check whether the stateDB needs to be updated.
If so, we should commit the pvtData to the stateDB.

As a peer can fail after committing pvtData of old blocks,
we might end up with a stateDB which is out of sync from
the pvtStore. Hence, this CR implements a recovery mechanism.

FAB-11806 #done

Change-Id: If6f8aa608080d7e02fb6c4236079814e3ed5c11f
Signed-off-by: senthil <cendhu@gmail.com>
Signed-off-by: manish <manish.sethi@gmail.com>