[FAB-13057] Remove applied index check in storage
For etcdraft chain, if there is applied index (>0) in
block metadata, then the node is expected not to be fresh
and should have wal data.

This CR removes this constraints because that's not the
case anymore with onboarding, since one node may pull blocks
from other nodes before bootstrap chain.

Change-Id: I678db4d1d71a24e08685d9c73eee7f4bc281771f
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>