[FAB-12729] Support subset of system channel OSNs
This change set changes etcdraft orderer node logic
to participate in a subset of the channels in the network.

Previously, before this change set - when an orderer can't find
its certificate in the channel config - it panics.

This is not a wanted behavior for application channels,
as we might want only a subset of the nodes that belong to the
system channel, to participate in a specific application channel.

After this change set, if an OSN finds it doesn't need to participate
in a channel - it simply presents to clients an "inactive channel".

This is needed, because channel creation transactions are classified
by whether the channel is known and registered in the registrar,
thus - if we are to not instantiate any channel object - some nodes
would act differently than other nodes with respect to classification
of target channel at the time of dispatching the transactions.

Change-Id: I254d270135fb4d9a105d04de893ae8ccdd13f0d7
Signed-off-by: yacovm <yacovm@il.ibm.com>