[FAB-13444] Prepare onboarding to multi-time use
This change set refactors the onboarding infrastructure to support
multiple use of it, in contrast to the current logic which
can only be used once.

Namely, it re-uses the current logic and introduces a new method
to the replicationInitiator:

replicateChains(lastConfigBlock *common.Block, chains []string)

Which forces replication of the given chain names, with the given
last config block.

The chains slice is passed to the cluster.Replicator as a filter
which prevents pulling chains that aren't among the chains slice.

Change-Id: I3331f2abb6a2879876644b2f5ef4ee48c4eb43fa
Signed-off-by: yacovm <yacovm@il.ibm.com>