[FAB-13843] Fix cclifecycle test flake
FAB-13471 Introduced a UT that does several (3) background listener updates.
The test only waits for the listener updates to fire, but there is a log invocation
that happens immediately afte the listener fires.

The next test that runs, overwrites the logger's reference and thus a data race
occurs (the previous test's spawned goroutine reads the logger's reference to write
the log, while the next test replaces the logger reference.

Fixed it by making the test wait until the log of the first test, logs
the last message that is logged during the listener dispatch - 3 times.

Ideally we'd have a logger instance for the production struct as a field,
but this solution is far easier and has no production code change, so its
surface area is smaller.

Change-Id: Ia7ef144ae8574783b2b7a4925f47b51e966f8942
Signed-off-by: yacovm <yacovm@il.ibm.com>