[FAB-14136] Always Deliver if cluster smaller than 3
When a single node etcdraft cluster is expanded and a new node
is added, the new node needs to pull blocks from the existing node.

However, the existing node loses leadership and then rejects deliver requests.

This change set, makes a raft leader not reject deliver responses if the cluster
has less than 3 members in it.

Change-Id: I75bd028d5a46fcb6ae81dc29012e3e839149c319
Signed-off-by: yacovm <yacovm@il.ibm.com>