[FAB-14463] wait for net error before healthcheck
After stopping CouchDB, the docker network artifacts may take some time
to cleanup. If a request is made to the CouchDB instance in that window,
the client will hang.

Attempt to work around it by hitting the CouchDB endpoint in the test
and waiting for a dial error to be returned. Once the dial error is hit,
proceed to the CouchDB down testing.

Change-Id: I40d2938956cb1bd1f015a23c2cba3a69a771141e
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>