rollback: fetch pvtdata of invalid tx (gossip)
we want to store pvtData of both valid and invalid
txs in the pvtData store. Till v1.3, we were storing
the pvtData of certain invalid txs only (which got
invalidated due to MVCC error). However, in v1.4,
we removed that too. Now, we need to store pvtData
of both valid and invalidTx so that during block
revalidation after a rollback, if an invalid tx
becomes valid, pvtData would be present and not
lost permanently from the network.

FAB-14310 #done

Change-Id: Id0996fabc72ad4c12535ad9f6dc546e4a9ce8131
Signed-off-by: senthil <cendhu@gmail.com>