[FAB-14607] Dynamically add channel verifiers
The channel verifiers for onboarding are used to verify blocks
that are pulled.

When the code was written, it was not possible to create channels
with a subset of OSNs in the channel, and therefore all OSNs that
were added to channels dynamically - loaded verifiers for these
channels at their startup.

However, now when we can create channels with subsets of OSNs,
we need the ability to dynamically register verifiers for channels
that are created.

This change set adds this capability in the least intrusive way-
before the blocks are pulled via dynamic onboarding, the verifiers
for channels that were discovered, are created on demand.

This change set also makes the block puller empty
its internal block buffer when Close() is called,
for additional safety.

Change-Id: Ife68ec1fe5b554a2089cc70640e4bc1d3014c39e
Signed-off-by: yacovm <yacovm@il.ibm.com>