FAB-14618 Store only nodeIDs in metadata
Presently, the block metadata encodes the TLS certificates of all of the
Raft consenters in the system for each block.  Because these TLS certs
are non-trivial in size, and there may be a large set of consenters,
this actually creates a significant amount of waste on the filesystem.

As a small optimization, this CR modifies the block metadata to only
store the nodeIDs instead of the full set of consenter info.  It then
correlates the consenter slice found in the channel config data with
this slice of nodeIDs to build a mapping between the two (which was
previously persisted).

Change-Id: Iaa66dacbcc48a041318c8a718099a873b9626240
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>