rollback: store pvtdata of both valid/invalid tx
we need to store pvtData of both valid and invalid
txs in the pvtData store. Till v1.3, we were storing
the pvtData of certain invalid txs only (which got
invalidated due to MVCC error). However, in v1.4,
we removed that too. Now, we need to store pvtData
of both valid and invalidTx so that during block
revalidation after a rollback, if an invalid tx
becomes valid, pvtData would be present and not lost
permanently from the network.

FAB-14311 #done

Change-Id: I0a5d7b00b9b8a8da4c487b59ff4fe237dfe7d260
Signed-off-by: senthil <cendhu@gmail.com>