FAB-14946 increase interval of proactive campaign
In proactive campaign, picked node starts campaign every TickInterval
till leader is present (either this node successfully claims leadership
or another node was already leader). This is excessive if network
latency is significant, and the term would keep advancing in extreme cases.
The interval of campaign should be increased to
            TickInterval * HeartbeatTimeout * 2

Change-Id: I6de195f28b0c259a626628ac3f8ae6a4a7c4571f
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>