FAB-15210 fix Raft UT flake
Recall that Raft reconfiguration is done with two rounds of consensus:
1) Config Block first
2) then Raft config entry

To precisely mimic leader failover between 1) and 2) in Raft UT, we
monitor number of Raft message being disseminated by leader, and
disconnect network at certain point of time.

However, in some cases, leader might optimistically replicate Raft
config entry along with the 2nd round of MsgApp for Config block
(recall 2 phase commit in Raft).

In such case, we need to pick the follower with higher index to be
elected as new leader, otherwise a node may never be able to claim
leadership and the test would fail.

In addition, this CR also reduces LeaderCheckInterval to 500ms to
speed up test run. It was set to 10s by default.

Change-Id: I7e735d5280928152420c174bfd2b1d54cbae4f25
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>