FAB-15174 Consensus migration: server boot sequence
Change the orderer server boot sequence to determine clusterType
from the system channel last config block or the bootstrap
block, depending on which block number is higher.

We do this because after migration the bootstrap file will have
ConsensusType.Type="kafka", whereas the system channel last config
block will have ConsensusType.Type="etcdraft". If clusterType is
not set to "etcdraft" at this point,  the creation of a raft
consenter will fail later on.

Change-Id: Ifb90f53562a748e7790b8036da0bb6b7a02ae87c
Signed-off-by: Yoav Tock <tock@il.ibm.com>