FAB-15175 Consensus migration: integration test
Cherry-pick for v1.4.2, as well as some code hygiene.

One test case of the migration flow on the kafka side. This
test case is dedicated to checking the filtering capabilities of
maintenance mode and the abort path.

One test case that executes the migration flow all the way from
a 3-orderer kafka system to a functioning 3-node raft cluster.

Change-Id: If7e61b2d9770e4d17f090472c71c44b365817ded
Signed-off-by: Yoav Tock <tock@il.ibm.com>