FAB-15829 Backport .gitattributes lineendings
Although fixed in master, many file types, like RST are being toggled
back and forth between LF and CRLF in release-1.4, this commit simply
adds those files to .gitattributes to let git handle it for us.

This is effectively a cherry-pick of FAB-14848 but because of the line
ending changes the cherry-pick merge conflicts were extensive, so
applied the .gitattributes only cleanly.

Change-Id: Ic0c7fa77c98549d2164b2d8633dfa42965cc3bde
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>