fix pvt and blockStore recovery for reset/rollback
As the ledger storage checks whether the height of pvtStore
and blockStore is equal during initialization, it results
in error during peer start after a rollback and reset. Note
that the error was random during peer start because the check
was made during initialization only when there was an ongoing
commit to ledger storage when the peer was killed for the
reset/rollback operation.

This CR makes the ledger storage initialization to accomodate
the reset and rollback effects as well.

FAB-16040

Change-Id: I0b9f044719306c4e71e9740b69f30048391d9b06
Signed-off-by: senthil <cendhu@gmail.com>