rollback: fix capability usage at gossip
Currently, we pass a static capability to the gossip which uses it
to check whether to store the pvtData of invalid tx. With a channel
update, the capability does not get updated. Hence, we need to use
the channel with bundle listener so that the capability gets
updated after a commit of a config update transaction.

FAB-16114 #done

Change-Id: I19d769432ce5783a1afdded5e7d4be9d682c2672
Signed-off-by: senthil <cendhu@gmail.com>
Signed-off-by: manish <manish.sethi@gmail.com>